Add actions for students and student selection

+++ resources/js/composition/useModelSelection.js
@@ -22,7 +22,11 @@ export default function useModelSelection (model) {
   }
   const selectNone = async () => {
     selection.value = []
-    selectedAll.value = false
+
+    if (selectedAll.value) {
+      selectedAll.value = false
+      return
+    }
 
     try {
       await $http.delete(`/selection/${model}`)
